Moved consts.hpp and added settings
Moved the file containing all the constants to the superfolder and added settings to it
