+++ catkin_ws/src/acr_core/src/consts.hpp
@@ -1,7 +1,8 @@
 #ifndef Consts_H
 #define Consts_H
 
-/* Identifying constants
+/* 
+ * Identifying constants
  * Identify parts by their id using a readable name
  */
 #define MODULE1 0
@@ -13,12 +14,20 @@
 #define U_FRONT_BOTTOM 2
 #define U_RIGHT 3
 
-/* Status constants
+/* 
+ * Status constants
  * Codings for various status states
  */
 #define MODULE_OK 1
 #define MODULE_FULL 0
 #define MODULE_IDLE 2
 #define MODULE_INTERACT 3
+
+/*
+ * Settings
+ */
+#define ULTRASONIC_MIN_DIST 50
+#define MOVE_SPEED 1
+#define TURN_SPEED 1
 
 #endif
