+++ catkin_ws/src/acr_core/src/core/states/state_concreteInteraction.cpp
@@ -31,6 +31,4 @@ void State_ConcreteInteraction::switchFrom() {
 	SensorData::sendModule(MODULE1, MODULE_IDLE);
 	SensorData::sendModule(MODULE2, MODULE_IDLE);
 	SensorData::sendModule(MODULE3, MODULE_IDLE);
-	SensorData::sendTwist(0.f, 0.f); // stop turning!
-
 }
